requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
free-proxy==1.1.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
loguru==0.7.2
lxml==5.1.0
mysql-connector==2.2.9
outcome==1.3.0.post0
peewee==3.17.1
pycountry==23.12.11
pycparser==2.21
PySocks==1.7.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
win32-setctime==1.1.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.2
